chore(core/modules): move perf manager to perf sub directory
