+++ apps/core/src/index.ts
@@ -13,7 +13,7 @@ import internalRoutes from './routes/internal';
 import { ProcessManager } from './modules/process.manager';
 import { GameManager } from './modules/game.manager';
 import { ConfigManager } from './modules/config.manager';
-import { perfManager } from './modules/perf.manager';
+import { perfManager } from './modules/perf/perf.manager';
 import { applyMigrations } from '@fxmanager/database';
 
 applyMigrations();
