+++ apps/core/src/modules/perf/perf.manager.ts
